pyproject.toml
[tool.poetry]
name = "habbit-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-rest-framework = "^0.1.0"
redis = "^5.0.0"
celery = "^5.3.4"
django-celery-beat = "^2.5.0"
django-filter = "^23.3"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
django-cors-headers = "^4.2.0"
coverage = "^7.3.1"
flake8 = "^6.1.0"
psycopg2-binary = "^2.9.8"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
